+++ src/routes/login.tsx
@@ -2,6 +2,7 @@ import { createFileRoute, Link, redirect, useNavigate } from "@tanstack/react-ro
 import { Eye, EyeOff, Loader2, Package } from "lucide-react";
 import { useState } from "react";
 
+import { PasswordStrength } from "@/components/atoms/PasswordStrength";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
@@ -15,41 +16,6 @@ import {
   translateAuthError,
 } from "@/lib/auth";
 import { supabase } from "@/lib/supabase";
-
-// ---------------------------------------------------------------------------
-// Password strength indicator
-// ---------------------------------------------------------------------------
-
-interface PasswordStrengthProps {
-  password: string;
-}
-
-const PasswordStrength = ({ password }: PasswordStrengthProps) => {
-  if (!password) return null;
-
-  const checks = {
-    length: password.length >= 8,
-    lower: /[a-z]/.test(password),
-    upper: /[A-Z]/.test(password),
-    number: /[0-9]/.test(password),
-    symbol: /[-_!?#@$%^&*()+={}[\]|;:'",.<>\\/~`]/.test(password),
-  };
-  const typeCount = [checks.lower, checks.upper, checks.number, checks.symbol].filter(
-    Boolean,
-  ).length;
-
-  return (
-    <ul className="mt-1 space-y-0.5 text-xs text-muted-foreground">
-      <li className={checks.length ? "text-green-600" : "text-destructive"}>
-        {checks.length ? "✓" : "✗"} 8文字以上
-      </li>
-      <li className={typeCount >= 3 ? "text-green-600" : "text-destructive"}>
-        {typeCount >= 3 ? "✓" : "✗"} 大文字・小文字・数字・記号のうち3種類以上（現在: {typeCount}
-        種類）
-      </li>
-    </ul>
-  );
-};
 
 // ---------------------------------------------------------------------------
 // Page component
